# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BR Tours and Travels API
  version: 0.3.0
  description: >-
    Express business API for the public website, authenticated administration,
    media, enquiries and notification operations. Booking requests are leads,
    not payments, issued inventory or guaranteed reservations.
servers:
  - url: /api/v1
...
paths:
...
  /pages/{slug}:
    get:
      tags: [Public content]
      summary: Published sanitized content page
      parameters: [{ $ref: "#/components/parameters/Slug" }]
      responses:
        "200": { description: Public page }
        "404": { $ref: "#/components/responses/Error" }
...
  /gallery/albums:
    get:
      tags: [Public content]
      summary: Published gallery albums
      parameters:
        - { in: query, name: destination, schema: { type: string } }
        - {
            in: query,
            name: package,
            description: Filter to albums associated through the package destinations,
            schema: { type: string },
          }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
      responses: { "200": { description: Public gallery album page } }
  /gallery/albums/{slug}:
    get:
      tags: [Public content]
      summary: Published album with public images
      parameters: [{ $ref: "#/components/parameters/Slug" }]
      responses:
        "200": { description: Public gallery album }
        "404": { $ref: "#/components/responses/Error" }
...
  /blog:
    get:
      tags: [Public content]
      summary: Search published articles
      parameters:
        - { in: query, name: q, schema: { type: string, maxLength: 120 } }
        - { in: query, name: category, schema: { type: string } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
      responses: { "200": { description: Published article page } }
  /blog/{slug}:
    get:
      tags: [Public content]
      summary: Published sanitized article
      parameters: [{ $ref: "#/components/parameters/Slug" }]
      responses:
        "200": { description: Public article }
        "404": { $ref: "#/components/responses/Error" }
  /faqs:
    get:
      tags: [Public content]
      summary: Published global or package FAQs
      parameters: [{ in: query, name: package, schema: { type: string } }]
      responses: { "200": { description: Ordered FAQs } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: br_admin_session
      description: Environment-configured opaque session cookie; never a browser-storage token.
  parameters:
...
    Slug:
      in: path
      name: slug
      required: true
      schema: { type: string, maxLength: 180 }
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, maximum: 1000, default: 1 }
    PageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 48, default: 12 }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            fields:
              {
                type: object,
                additionalProperties: { type: array, items: { type: string } },
              }
            requestId: { type: string }
  responses:
    Error:
      description: Safe API error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }